[change] server modify
