--- src/main/java/com/cxuy/handler/WorldHandler.java
@@ -1,27 +1,0 @@
-/*
- * Copyright (c) 2025 liangbeiyuan.
- * Licensed under the MIT License. See LICENSE file in the project root for full license information.
- */
-
-package com.cxuy.handler;
-
-import com.cxuy.framework.util.JsonUtil;
-import com.cxuy.framework.util.Logger;
-import com.cxuy.server.SimpleHttpServer;
-import com.cxuy.server.annotation.HandlerMethod;
-import com.cxuy.server.annotation.HttpHandler;
-import com.cxuy.server.annotation.RequestBody;
-
-@HttpHandler(path = "/helloworld")
-public class WorldHandler extends SimpleHttpServer.Handler {
-
-    @HandlerMethod
-    public void post(String name, @RequestBody Body body) {
-        Logger.d("WorldHandler", "receive " + name + ", Body=" + JsonUtil.toJson(body));
-    }
-
-    public static class Body {
-        String name;
-        int id;
-    }
-}
